BIOM datasets are developed more
-- HDF5 support
Do not depend on any module but instead a transform:
https://github.com/VEuPathDB/EuPathGalaxy/commit/d78a6489d5
to a made up format with sample/taxon details in JSON and data in TSV

Convoluted and inefficient but kind of work.
This was done because it was difficult to add depedencies to the installer.

-- Write to the DB with sqlldr
Over ten times faster than the sequential inserts.

-- Logging
Require Log4perl. We have it on `elm` so could be noncontroversial.

-- Other
Minor development for lineages and sample details modules
Longer columns for lineages
DB indexes where I've seen slowness