Allow NCBITaxons from sres.ontologyterm table
The problem was that the objects `_ontology_term_to_identifiers` and `_ontology_term_to_names hashes`
were used to keep two kinds of data - taxonomy info, and content from the .owls.

This did not work when an entry in the .owl needs to be used to make a Sres::OntologyTerm, but bears an NCBITaxon ID.

I've split the hashes, changed the behaviour to populate them with two queries, and used the correct source in each place.
Now the places that do $taxonId = ... instead of $ontologyTermId now access the new hash.